Proper pointer-to-jlong casting

+++ native/sdljoystick4java/com_illcode_sdljoystick4java_Native.cpp
@@ -4,5 +4,5 @@
 #include "com_illcode_sdljoystick4java_Native.h"
 
 JNIEXPORT jlong JNICALL Java_com_illcode_sdljoystick4java_Native_getDirectByteBufferAddress(JNIEnv * env, jclass klass, jobject obj) {
-	return 0L;
+	return (jlong)(intptr_t) 0L;
 }
